Just move weather structs into a specific file: WeatherModel
