+++ Cab share/LocationManager.swift
